requirements.txt
pandas==1.4.2
psycopg2==2.9.3
python-decouple==3.6
requests==2.27.1
SQLAlchemy==1.4.36